requirements.txt
Flask==3.0.0
replicate==0.25.0
python-dotenv==1.0.0
Werkzeug==3.0.0
gunicorn==21.2.0
